Consolidate the search space into search/: space.py + features.py, the PLACE pin
Phase 1 of the knob-search restoration:

- search/space.py is the ONE search-space file: every Knob declaration
  (moved from pipeline/forks.py, deleted) plus the enumeration value
  grids (moved from lowering/tile/_catalog.py, deleted). Help strings no
  longer name the retired 020_schedule pass.
- search/features.py holds every featurizer (knob_features,
  tile_signature, the D_*/MMA_* encodings, masked_axis_features), moved
  out of knob.py; knob.py shrinks to pure infra (descriptor, registry,
  env plumbing, axis-key helpers, display).
- The pin-only PLACE knob: one structural-placement family
  (PLACE@cone / PLACE@fold / PLACE@tuple = auto|fuse|cut, precedence
  element > bare > built-in auto). Knob.narrow_at() is the per-element
  env read; parse_knob_spec canonicalizes case per key part (family
  upper, element lower) so PLACE@fold / TILE@d survive the aggregate.
  place_decision() resolves the vocabulary; auto keeps today's emission
  (fuse everywhere). PLACE@fold=cut is the multi-kernel attention
  escape, PLACE@tuple=cut the two-pass softmax stats; a forced fuse on
  an uncertifiable kernel degrades to cut with a log line. This retires
  the dead DEPLODOCK_FLASH spelling (nothing read it since the
  recognizer merge).
- _MAX_BLOCK_THREADS deduped (space.MAX_BLOCK_THREADS, one constant).
- Docs: pipeline/ARCHITECTURE.md modules + knob sections rewritten
  (declarations in search/space.py, PLACE row replaces the stale
  CUT/FLASH/ONLINE_SOFTMAX rows); every 020_schedule reference now
  names the _schedule helper inside 010_recognize; fork.py's dead
  085_warp_specialize references dropped.
